Daily progress.
Added support to DeviceAgentVector, for all insert/erase ops.

This is not currently working, i did not account for the impact of initUnboundBuffers() being called after an erase/insert action had already been applied to the underlying AgentVector.

This appears to now be fixed for inserts, but erase still needs fixing.

I'm half in mind, to make an earlier change to track the known size to prevent this 'edge' case, rather than complicating everything else to work around it.