Started working on an example for demoining visualisations.
Found several bugs in current visualiser.

Detailed in visualiser repo issue 35, a visassert is thrown when I try to render 2 agents, if i comment out that debug check, a diff visassert is thrown because an agent lacks Z coord, after fixing that that the 1st agent renders incorrectly, only moving in the Y axis.